requirements.txt
azure-cognitiveservices-speech
streamlit
openai
elasticsearch
elasticsearch-llm-cache